refactor: extract AppErrorSubPresenters.swift from AppErrorPresenter.swift
Extracts 2 domain-specific presenter classes (TranscriptPersistenceFailurePresenter,
PostTranscriptionFailurePresenter) from AppErrorPresenter.swift into a new
AppErrorSubPresenters.swift.

Byte-identical move; SHA-256 024ef5e6bfbc62c9055f2b42fd8579a5ec9b68c29e9ede168c3e13067c3e947e
verified against origin/main:AppErrorPresenter.swift lines 29-91.

Applied memory rule "watch for annotations on line immediately preceding
deleted blocks": verified @MainActor at line 92 (for the primary AppErrorPresenter
class) is preserved before deleting the sub-presenter range 28-91.

Reduces AppErrorPresenter.swift from 310 to 247 lines.

Tests: 667 (unchanged).
